updated singledot source analysis for dots 3,4,5 with accev
This singledot analysis is part of my investigations into accumulated evidence.
For that purpose I have included the accev regressor for the 3rd dot only which
consists of the sum of the first 2 dots. To avoid collinearities I therefore had
to exclude the first 2 dots from the analysis. I included specifically accev_3,
because that is the first accumulated evidence that is not directly related to
dot_x, because it is a sum of two dot_xs. Also it's early enough so that it may
not be related to the executed response. Note that the response regressor is
still correlated with the dot_x regressors, but only with coefficients below
0.4. The condition number of the design matrix also was around 10 when I
included abs_dot_x and abs_dot_y, but it dropped to values around 5, when I
excluded the two regressors. This is a bit strange as the correlations between
response and dot_x were actually larger than the correlations with abs_dot_x and
_y.

I updated the analysis to include the new normalisation of the design matrix and
the previously forgotten normalisation of the data.

Results will be available in:
	source_singledot_201709071126.h5