including dotx_time into complete-trial inference with delay 0.3
I already know that accumulated evidence for a dot is correlated with the
x-coordinate of that dot, but how severe is the issue and does it influence the
results for the accumulated evidence regressor in the complete-trial analysis
across all time points? So I created the dotx_time regressor in analogy to the
accev_time regressor with the difference that after the response dotx_time is 0
instead of maintaining the last value. Because only in about 30% of all trials
the response is before the end of my analysis time window, I presume that this
anyway has no big influence.

Results are stored in
	source_time_201709081726.h5

I observe two rather unexpected effects:
	1. results for accev_time resemble that of accev_time without dot_x and
	   a delay of 0.4
	2. the only significant effects of dotx_time are in left ACC (a24, p32),
	   left IFJa and left AVI (insular cortex)

Stefan suggested before that the GLM, when it contains collinear regressors,
automatically represents the orthogonal information between the regressors. I
don't understand fully how that works, but at least for 1. this makes intuitive
sense: I have seen before that accev orthogonalised with respect to dot_x is
equal to accumulated evidence for the previous dot. This, again, is equal to
delaying accev_time by a further 100 ms in my analysis and I see the
corresponding result.

For 2. I'm not sure at all what is going on. When I ignore the significance
threshold I see, weakly, many areas that I know from previous analyses of dot_x
such as visual areas, posterior cingulate, parietal and motor areas, but it's
unclear why I suddenly see most significant effects in these left frontal areas.
Is this the difference between accev and dot_x? How could dot_x be represented
in frontal areas, but not accev?

What this result shows is that accev_time is heavily influenced by the inclusion
of dotx_time into the analysis. This is the case even though the condition
numbers of the design matrix are well below 10. I should really check this
again. The correlation between dotx_time and accev_time is around 0.61 for all
subjects. The correlations between accev_time and motoresponse are between 0.38
and 0.63 across subjects with slowly responding subjects tending to have
smaller correlations.