visualise the correlation underlying dot_x effects in motor cortex
This notebook investigates the correlations underlying a particular regression
result. The idea is to gain confidence in a potential linear relationship
between a numerical regressor and the data. Specifically, we want to exclude the
possibility that a correlation is only driven by particular subsets of the data.
The difficulty in this endeavour is the small size of the effect in relation to
the data standard deviation. This means that it is extremely difficult to see
any relationship in a simple scatter plot. The notebook provides a series of
correlation plots finishing with one which only shows the correlation line above
means of data bins with their standard error. As this allows to draw the
correlation in the scale of the effect and not in the scale of the data standard
deviation, the effect can be seen most clearly.

I did this to particularly investigate the representatio of dot_x in the motor
cortex and correspondingly have chosen to investigate left primary motor cortex
420 ms after dot onset, because this is when I saw the strongest motor cortex
effect.

I looked at the correlations for dots [1], [1,2], [1,2,3], [1,2,3,4,5] and saw
clear (but very small) effects. These effects appear to be driven by a true
linear relationship between dot_x and the motor cortex signal. At least I did
not see any evidence that the effect is driven by extreme dot_x values. Also I
do not see a clear step from negative to positive dot_x values, but more or less
a gradual and somewhat noisy rise in signal with a rise in dot_x.