compare correlations between regressors and source signal across regressors
I wanted to see whether I can reasonably argue that the signal in motor cortex
at 420 ms better corresponds to accumulated evidence than only the x-coordinate.
My idea was that the correlation should be stronger for accumulated evidence in
that case. Although the correlation with sum_dot_x as accumulated evidence is
indeed higher than that with dot_x, the difference is only small and actually
disappears when I look at the first 5 dots together.

I find it quite remarkable how close the means of the bins for dot_x and
sum_dot_x are in the linear model plot. Well, it is not surprising that their
x-values are similar, because they just represent quantile values. When the
x-values differ, this just means that the density of the regressor values
differs slightly in relation to the original regressor standard deviation. But
also the y-values, i.e., the average signal in left primary motor cortex within
the regressor bins, are surprisingly similar in the sense that the y-values for
sum_dot_x tend to follow the up and down of the y-values for dot_x. This means
that the same trials tend to fall into the bins for sum_dot_x and dot_x.

Perhaps, this is only a first impression, however. As I checked the plot for
individual dots from 1 to 5 I saw reasonable variance and differences in
y-values and it was hard to see consistent deviations across the dots. Perhaps
the best visualisation for seeing consistent deviations is the linear model plot
for all 5 first dots as it is like an average across the dots. This plot shows
that the dot_x bin means resemble a sigmoid which flattens at the bottom and the
top. The sum_dot_x bin means, on the other hand, appear to flatten only at the
bottom while the mean of the rightmost bin extends well above the regression
line. Given that the left motor cortex supposedly preferentially encodes
movements with the right half of the body and positive sum_dot_x values encode
evidence for a response with the right hand, the overall direction of the
correlation makes a lot of sense, but what does it mean that the largest
sum_dot_x values average to a comparably higher signal than those of dot_x?

The comparable overshoot of the rightmost bin is there for all dots except for
the 5th, but I got the impression that there is something strange going on in
the 2nd and 3rd rightmost bins of the 5th dot - it appears that trials with high
signal values have moved from the 3rd rightmost bin of dot_x to the 2nd
rightmost bin of sum_dot_x. This could be caused by our manipulations of dot_x
for the 5th dot which include comparably large x-positions that could overwrite
a lot of previously accumulated evidence (the maximum frequent deflection of
x-positions in the 5th dot is (+/-)180 while it is about 120 in the 4th dot and
the sum_dot_x values of the 4th dot are in the range from -260 to 260 with most
being within -120 to 120). Actually, you should think about it the other way
around: the accumulated evidence from the 4th dot will shift the x-positions of
the 5th. The high values moved from the 3rd to the 2nd rightmost bin, because
the x-positions of the 5th dot come in chunks and also sum_dot_x of the 4th dot
has two clear peaks in the density. So a large chunk of data of dot_x moved to
higher sum_dot_x values relative to the standard deviation of sum_dot_x, thereby
crossing the bin boundaries. As everything is symmetric, the same should have
happened for the negative regressor values. Perhaps this is the case, but it's
harder to spot, because there is not such a big difference in bin means for the
bins on the left. Anyway, the only thing this shows is that you get into
problems with the binning of the regressors, when you look at the 5th dot in
isolation, because of our very discrete manipulation of 5th dot x-positions. It
should not be a problem to include the 5th dot in an analysis with other dots,
though, because then the distributions of dot_x and sum_dot_x become much
smoother, although they still exhibit some recognisable peaks due to our choice
of dot x-positions for the experiment.

practical change: I preload data for the first 5 dots now, but allow analysis
for a subset of them. This allows faster interaction and exploring of effects.