correlation plots for the paper
Based on the notebook I have created a cleaner script for generating two
correlation figures for the paper. The first shows the correlation of dot_x with
both primary motor cortices for 4 different time points. In the second I compare
the correlations for three different regressors (dot_x, sum_dot_x and
motoresp_lin).

The new motoresp_lin regressor represents a linear increase or decrease reaching
a common level at response time. It increases for right choices and decreases
for left choices. The regressor is a simplification of motoresp which
additionally has different phases (0, linear increase from 600 ms before
response, slightly slower increase from 300 ms before response, linear drop
after response) while motoresp_lin is really just a simple linear function of
time. To avoid that motoresp_lin switches sign at some point before the response
I defined it such that it reaches 0 at the maximum RT allowed in the
experiment (2.5 s) before the response. Note that timeouts are encoded with
RT=5. Timeouts may, therefore, lead to sign flips of the motoresp_lin values and
should probably be excluded from analysis. Timeouts should be relatively rare.
So they shouldn't affect the analysis too much, but I'd rather check this.

To include the time-based motoresp regressor in an analysis in which I pool data
for different dots I used the fact that for a given dot onset time different
dots correspond to different times within the trial. So I associated, e.g., 120
ms past dot onset for dots 1 to 5 with trial times 120, 220, 320, 420 and 520.
This means that the 5 dots are associated with the same data as the motoresp
regressor at the corresponding time points.