investigating results of single dot analysis in source space
The main motivation for the single dot analysis in
	source_singledot_201709071126.h5
was to see the correlations with accumulated evidence at an early time point at
which I can theoretically exclude that the correlation is driven by motor
preparation. For example, I see strong effects for accev_3, i.e., the sum of the
x-coordinates of the first two dots, already at 320 ms. These effects involve
primary motor cortex and are so early that they should not relate to motor
preparation. The largest effects, however, for accev_3 occur at 420 ms with
three strong hubs in posterior cingulate, somatosensory and motor cortex and in
anterior cingulate. There are also effects in operculum and insula, especially
on the right.

When plotting the sum of significant mlog10p values across times for accev_3,
you'll see peaks at 170, 220, 320 and 420 ms. There is only a small hillock at
120 ms. These times reflect the important events that I saw in the sequential
analysis. For example, at 220 ms I observe strong effects in V1 which can be
attributed to the early sensory processing of the second dot 120 ms after its
onset, but notice that there are also effects, e.g., in mid-cingulate at this
time. 320 ms is supposedly the time when the first dot is processed for
inclusion into accumulated evidence leading to strong effects in accev_3 and 420
ms is the same for the second dot giving the strongest effects af accev_3 across
the brain. Two characteristic drops in the sum of mlog10p values follow at 500
and 600 ms which probably reflect the corresponding drops in the late dot_x
effects at 500 ms after dot onset.

The other main purpose of the analysis is that I can see correlations of the
response throughout the trial. To my surprise there are significant effects of
response before 700 ms. Specifically, there are strong effects around 320 ms and
400 to 460 ms. Smaller effects (in sum of mlog10p values over all areas with
significant effects) also occur at 220, 260, 520 and 620 ms. The largest effects
happen after 700 ms and probably represent the actual motor preparation.

So what are these early effects? One striking feature is that around 220, 260,
320 and 420 ms always V6 is involved. Also areas 7m and 31pd from posterior
cingulate cortex show effects at many of these time points. The most early
effects occur at 320 ms mostly in posterior cingulate, but also in left
operculum and insula (FOP2-4, Ig, MI). The effects at 620 ms are entirely driven
by V1 and v23ab. Because the 120 ms effect of the 6th dot occurs at 620 ms, this
suggests that often people used the 6th dot to make their decision.

Why do these areas, such as V6, correlate specifically with the binary response,
i.e., the final choice of the subjects? Right V6 shows the most consistent
activations in the sequential analysis, i.e., it has effects at the largest
number of time points across all areas, but this is not true for left V6 while
for the response V6 is often activated bilaterally. Perhaps the way V6 encodes
accumulated evidence is more like an all or nothing instead of a graded
encoding, but why should this happen specifically at 220, 260 and 320 ms? dot_x
does not correlate with the V6 signal at 120 ms after dot onset. So the 220 ms
effect cannot be explained by the early 120 ms effect of the 2nd dot. It could
also be that trial-by-trial fluctuations in V6 that are at first unrelated to
the stimulus lead to preferential processing of the stimulus so that it
influences the decision outcome.

Another alternative is that it's hard to interpret these results, because they
interact with those for the accumulated evidence after the first two dots: What
I see could only be difference effects that were not explained by accev_3
directly. Under this assumption an analysis without accev_3 should exhibit
stronger early effects of response. Similarly, there could be an interaction
with dot_x (although probably less so). For example, I may only see the clear V1
effects in response at 620 ms, because I don't have a dot_x_6 regressor.

Whatever these early response effects are, they are not motor related: Actual
motor or premotor areas only show effects from 700 ms onwards. The caveat here
being that I may not see the motor areas, because accev_3 has already explained
them away.