checking whether it would make sense to only use error trials
The idea was to reduce correlations between accumulated evidence and response by
concentrating on trials in which subjects chose apparently not based on
accumulated evidence. So I identified error trials as those in which subjects
chose a response that was opposite of what the accumulated evidence indicated at
a particular time point (here: after the 4th dot). In these trials there
shouldn't be a relation between response and accumulated evidence, but of course
there is: By construction there is a very strong anti-correlation (-0.8) between
the accumulated evidence of dot 4 and the response. Only accumulated evidence of
dot 2, 6 and 7 have pretty low correlations with the response for error trials.
While it is no problem for positive, categorical regressors to be
anti-correlated (one is 1 whenever the other is 0), anti-correlation of a proper
parametric regressor is as bad as correlation. So it makes no sense really to
restrict the analysis to error trials.

For completion I have now also added plots showing how the dot_x regressors
themselves correlate with the response. Apparently the 4th dot is least
correlated with the final response (~0.15) while the 5th (~0.39) and 8th (~0.38)
dots correlate most with the response. This is surprising in two ways: i) Why is
the 4th dot so unrelated to the response? ii) Why is the 8th dot still
correlated with the response?

i) may be an artefact of our choice of stimuli: Perhaps the 'long' trials that
Hame chose to ensure that people see the 5th dot have very uninformative 4th
dots, or even contradicting 4th dots, i.e., whose x-coordinates are opposite of
the correct side.

ii) is harder to explain. The median response time is around 900 ms which means
that the onset of the 8th dot is only 200 ms before the median response time and
I didn't expect so late dots to have an impact on the response. On the other
hand, this also means that in many trials the 8th dot will still be processed,
perhaps even in the phase starting from 300 ms past dot onset.

The 9th and 10th dot have smaller correlations with response (~0.33) and from
the 11th dot correlations are low (~0.23). I guess that this low level of
correlations around 0.2 is the correlation that should be expected simply
because the drawn dots will on average point into the direction of the correct
decision. Then this also means that subjects sometimes still use up to the 10th
dot to make their decisions. This is also contained in the accumulated evidence
correlations: They increase roughly up to the 10th dot and then stay level at
about 0.68.