updated dot_x brain area figure using absmean beta estimates
Using my developments of color handling directly in PySurfer I have updated the
color displays on the inflated surface. Also, I have now switched to using
directly the simple second-level statistics of the original inference files as
there are almost no differences to the more complicated folded-normal
second-level inference.

I have checked 'absmean' and 'abstval' for display of dot_x results. The figures
are almost the same with only small differences. This is probably also helped by
me showing always the same brain areas (significant clusters) and small
differences in colour are not so easy to spot. I go with absmean for now, as it
is the more meaningful quantity.