GP-based analysis of relation between source signals and evidence and motor prep
I replaced the previous figure with a simple linear regression fit with one
showing a nonlinear Gaussian process fit. Also I now produce a figure which
estimates the source signal aligned to response time in time coordinates. This
is easy using the GP fit of the motor preparation signal, because that is only a
linear transform of time relative to the response.

For GP fitting I use sparse GP regression of GPy. With the 15 inducing inputs I
use, the optimisation of the GP hyperparameters (including the inducing inputs)
takes only a few seconds. So figures can still be produced on the fly.