code for mapping HCP atlas from HCP to fsaverage geometry
I originally just used the HCPMMP1 atlas provided in MNE which they got from

https://figshare.com/articles/HCP-MMP1_0_projected_on_fsaverage/3498446/2

In the meantime Glasser et al. updated the instructions for mapping between
their standard brain and fsaverage. I re-mapped the atlas using the new
instructions, see map_HCPMMP1.sh. For comparison I also implemented the old
mapping in map_HCPMMP1_old.sh, but I haven't actually compared the two sets of
labels.

My original motivation for doing this was to learn how to map between the HCP
standard brain and fsaverage in general, because I wanted to map a set of
borders in the motor cortex defining where regions of the body parts are. My
first attempt at that is stored in map_motor_labels.sh, but this does not work
probably because I cannot properly transform the borders to gifti-labels.

For information I also commit the pdf-instructions provided by Glasser et al..

Requirements for the mappings are:
- the downloaded HCP data files
- the connectome-workbench
  https://www.humanconnectome.org/software/connectome-workbench
- pre-loaded freesurfer
  source fshomedir/SetUpFreeSurfer.sh
- HCP standard meshes available at
  https://github.com/Washington-University/Pipelines/tree/master/global/templates/standard_mesh_atlases