fixed mapping of motor borders and clean-up of mapping scripts
Based on a response from the HCP-Users email list from 27.10.2017 I could
complete the mapping of the motor subregion borders. The basic problem was that
the transformation of borders to ROIs mapped each border to a separate ROI in a
column of a data file. The ROIs are separate because they could overlap which
they don't in my case. Labels should be represented as a single vector of
(index-)integers. The email gave the commands that achieve this transformation.

There was a warning that there could be something going wrong, if the borders
are not fully closed and have gaps, but checking the generated labels
(tmp.label.gii) against the original borders on the surface I found that they
overlap very well.

For the transformation into a freesurfer annotation I also had to assign a
unique color to each label, because freesurfer uses the colors as index.