new vertex-level inference in premotor and motor cortex
To find futher evidence for involvement of primary motor cortex in decision
making I investigated the precise locations of correlations between the signal
in individual source vertices and evidence. To limit computation time I only
investigate the time from 300 to 490 ms after dot onsets (first 5 dots).
Otherwise the procedure is the same as before.

results are in:
	source_epochs_allsubs_HCPMMP1_201710231731.h5
	source_sequential_201710231824.h5