visualisation of vertex-level signal in premotor and motor areas
In this visualisation I overlay the t-values of the premotor and motor vertices
on top of the area borders so that it is possible to interpret the loci of
highest correlation with respect to these areas. I also show the motor regions
associated with the eyes and upper extremities including the hands as a
cross-check for the plausibility of the result.

The code for this visualisation is based on show_VEF_sources, but generalises
it so that it can be easily reused for different new sets of vertices. Also I
make use of the new pysurfer color routines.