make movie from images of motor effects
generated movie doesn't show well in built-in video player, but looks good in
VLC

Effects in the left hemisphere tend to be weaker than in the right, but tend
to stay in eye and upper extremity regions. Effects in the right hemisphere
start out pretty low in eye and face regions, but wander upwards towards hand
regions by 400 ms.