all vs. early data comparison of correlation with dot_x in motor cortex
Stefan thought that the previous comparison of different correlations in
primary motor cortex was hard to interpret. Indeed, the normalisation of
regressors meant that it was not necessarily clear what exactly was compared.
As we are anyway shifting more away from accumulated evidence, I decided to
combine the figures for all and only early data into one and only show
unnormalised dot_x correlations while keeping the motor response figure.

It turns out that when the regressors are not normalised the x-coordinate
correlations look clearer when all trials are included and not only those in
which the response was to the contralateral side.