repeat motor vertices analysis with exclusion of late data + accev
Stefan thought that it was important to exclude the possibility that the dot_x
correlations we see in primary motor cortex are mostly related to the motor
response. This just repeats my analysis of the correlation plots with the
greater resolution of the individual motor vertices. Specifically, I have
re-implemented the rt_thresh parameter so that it excludes data that is after
RT-rt_thresh (with rt_thresh=500 ms here). Non-surprisingly I find similar
effect patterns as for the full analysis, but which are much weaker
statistically. To support this comparison I now scale the colormap so that the
t-values which are deemed statistically significant (FDR corrected with
alpha=0.01) are mapped to the fmid color.

Results
-------
of the GLM inference are stored in
	source_sequential_201711031951.h5

Instead of t-values close to 9, the t-values stay below 6.5 when late data is
excluded. Also, the effects are much more patchy with spurious occurrence and
disappearance of effects. There are almost no effects in the right hemisphere
from 410 ms onwards for dot_x while the left hemisphere shows the remnants of
the effects from the full data for most time points from 400 ms, but nothing
before 400 ms.

Effects of accumulated evidence are very similar to those of dot_x when
adjusting for the time appropriately (accev at 300 ms corresponds to dot_x at
400 ms). For late data excluded this means that accev at 300 ms shows
reasonably strong effects in primary motor cortex at the transition between
eye to hand areas in both hemispheres which then reduce until they are pretty
much non-existent around 350 ms. There is a weaker shadow of this at 400 ms,
but no further effects after that. For late data included, t-values for accev
are slightly larger than for dot_x, but the patterns appear similar except that
they are more comprehensive (include more vertices) and are flipped in sign, as
expected. The average effect size decreases for accev from 300 ms onwards.
While by 490 ms only faint effects are left in the left hemisphere, the right
still shows relatively strong effects.

Compared to my previous color scaling which highlights the 5% largest absolute
t-values the new significance-based scaling leads to spatially much larger
effects for the results including (almost) all data. The effect patterns
consist of long stripes covering primary motor cortex, area 3 (a and b in
one stripe) and areas 1 and 2 (in one stripe). A similar pattern exists in
pre-motor areas, but there the pattern is wider. The stripes mostly cover the
hand region, but for areas 4 and 3 also cover eye and face regions. In the
right hemisphere, the verteces with the largest effect in areas 3 and 4 appear
to be in the eye region, or at the border between hand and eye. In the left
hemisphere there is a gap between effects in area 4 eye region and area 4 hand
region and I cannot tell from the pictures which has the largest effect size.

Issues
------
- are fewer effects for exclusion of late data due to smaller statistical
  power, or because they are really gone? really informative are only the
  effects that are still there (not many)
- exclusion of late data affects subjects differently: those with fast RTs will
  have fewer data points in the analysis, also data from 5th dot will be much
  more frequently excluded than data from earlier dots. This can lead to
  various asymmetries that may invalidate the statistical measures, e.g., the
  summary statistics approach assumes equal variance across subjects which then
  cannot be ascertained