demo of the task as run in the MEG
I made this in preparation of my talk in the groupmeeting and also to get a
feeling for how likely it is that people actually tracked the dot with their
eyes by actually moving them.

The distance between the targets and the standard deviation of the dot movement
is chosen so that it replicates the setting in the experiment. To do this
faithfully the dpi of the screen has to be known. For now I don't set the size
of the shown symbols in relation to the dpi of the screen, because I don't know
whether matplotlib does that automatically.