investigate response-related effects in motor vertices
I wondered how the x-coordinate effects differed from those that can actually
be attributed to button presses. My first approach was to simply run a
singledot analysis for the motor vertex sources. This showed the largest
effects around 800 ms, as I could expect from the sensor-level average response
effects, but strangely the effects were almost exclusively in the right
hemisphere.

This analysis also included regressors for the x-coordinates of dots 5, 6 and 7
which were supposed to exhibit effects in the considered time window starting
at 700 ms after first dot onset. Because these regressors are correlated
(r~0.3) with the response regressor, it could be that they at least partially
explain away response-related effects. For the 5th dot, for example, I see a
weaker representation of the typical 400 ms past dot onset effect at 800 ms,
but only in the left hemisphere. This fits to the interpretation that the
x-coordinate regressors share effects with the response regressor and
(arbitrarily?) divide up partial effects between them.

To overcome this alignment between dot effects and the response effects I
finally implemented an analysis which investigates response-aligned effects. I
assume that the responses are suitably jittered so that dot effects are
averaged out and correspondingly left out dot regressors from the analysis. The
major issue with this analysis is that it is still based on the original epochs
so that not in all trials data from the time of the response is available.
Specifically, for the time of the response the median number of trials
available per subject is only 146 with 5 subjects even having less than 30
trials, because they had generally long RTs. Nevertheless, I observe strong,
bilateral response effects in this analysis. Interestingly, following my
previous correlation (figure) analysis on the level of brain areas, I see the
largest effects 30 ms after the response. What I didn't expect is that there
are only weak, or no effects building up to the response (e.g. at 50 ms before
the response).

The major finding of this analysis is that the response effects, i.e., increase
in source currents for one response with a decrease in source currents for the
other, mirror the 400 ms dot onset aligned x-coordinate effects. This includes
effects in the eye region. Specifically, it appears that eye region effects in
primary motor cortex precede effects in the hand region, although effects in
the hand region of somatosensory cortices and in premotor areas occur together
with the early eye effects.

The interpretation of these findings is that either a) button presses are
accompanied by eye movements to the corresponding side, b) source localisation
is not precise enough to discriminate the eye and hand regions properly, c) the
preparation and execution of hand movements is in this task coupled to
preparation of eye movements even though they may not be executed, or d)
something else (could subcortical effects be somehow shadowed as eye effects in
cortex in our source reconstruction?).

Results are in:
	source_singledot_201711061631.h5 (dot-aligned, 700 to 890 ms)
	source_singledot_201711081136.h5 (response-aligned, -200 to 50 ms)