investigating build-up in primary motor cortex from trial onset
Given that the effects in primary motor cortex look very similar at 400 ms from
dot onset and at the time of the response, my current hypothesis about the
involvement of primary motor cortex in perceptual decision making is that it
pre-activates movements proportional to the current evidence. The question
remains why nobody reported this before. Perhaps they simply attributed any
build-up in primary motor cortex to unspecific motor preparation. But then I
also wonder why I don't actually see any slow build-up in the response aligned
analysis, but only effects within 50 ms from the response. Also I wondered,
when primary motor cortex activates proportional to (accumulated) evidence,
shouldn't I see a general build-up of currents starting from first dot onset?
Even if (accumulated) evidence is only represented from about 300 ms after dot
onset, on average there should be a continuous representation of evidence from
300 ms after first dot onset. Under the assumption that evidence for a left
decision on average gradually builds up towards a left choice, the build-up
should be seen in the contralateral (right) primary motor cortex.

So in the present analysis I try to estimate the timecourse of the average
motor currents prior to a contralateral choice starting from trial / first dot
onset. I would expect that this starts to rise from 300 ms, because this is
when I supposedly first see representations of evidence in primary motor
cortex. I use sparse GPs to denoise the time courses from individual trials
within subjects and then compute a weighted average across subjects.

There is build-up, but it doesn't appear to start at 300 ms. In the left
hemisphere the overall mean is pretty flat until a rise starts around 400 ms.
In the right hemisphere the overall mean drops until about 200 ms after which
it rises to a similar initial level as in the left hemisphere (reaching that
level around 330 ms) and then rises further from about 400 ms, but less
strongly than in the left hemisphere. Given that the largest dot-centred
effects of evidence in primary motor cortex are around 400 ms, the start of the
rise at 400 ms appears to be consistent with my considerations with the
limitation that I wonder why the smaller effects between 300 and 400 ms are not
represented here.

A complication of this analysis is that I only observe 300 ms of the build-up
and the end of this period is already the start of when I see average response
effects (at 700 ms). This also means that this build-up is not necessarily a
direct effect of the response. The response effects should come on top of this
build-up. My prediction therefore is that from 700 ms there should be a visible
increase in the slope of the build-up. I should re-run the analysis for the
complete time window up to 890 ms to find out.

One thing to keep in mind is that this analysis is very closely related to the
correlation of the motoresplin regressor with the signal in primary motor
currents (done in a previous version of make_correlation_figures.py).