mean timecourse analysis until 890 ms, uncertainty over mean prediction
Following up on my last commit I repeated the analysis for the full time
window. In contrast to my prediction there is no increase in slope of the
build-up after 700 ms. Rather, the average signal in the hemisphere
contralateral to the choice peaks around 750 ms and drops with a similar slope
as the build-up. So it is not possible to separate contributions of
evidence-proportional motor preparation from purely motor-related response
effects. This is because the response effects are smeared over large periods of
time due to the jitter in response times.

On the other hand, I don't really understand why the average signal drops
already after 750 ms, when 60% of all RTs are slower than 1 s and the actual
response-related effects only occur within 50 ms of the response.

I will double-check whether this is an artefact of estimating this with GPs,
but the new results for the whole time window are entirely consistent with the
previous analysis only up to 700 ms, i.e., the estimated time courses are
extremely similar between 0 and 700 ms.

uncertainty over mean prediction
--------------------------------
Instead of using a weighted average I switched to using the mean prediction.
The mean time courses of the two approaches are very similar, but I can easily
compute the standard deviation of the mean prediction as this is related to a
sum of Gaussian random variables (divided by N) whose variance is simply the
sum of variances of the individual variables.