compare GP vs simple average, summary statistics uncertainty
To check whether the GP-based analysis gives reasonable results I computed an
equivalent analysis by simply averaging the corresponding data at the given
time points. The result is much noisier than the GP result and the GP noiseless
prediction does indeed look like the underlying noise-removed, smooth time
course. So the GP analysis appears very reasonable.

Alternative to the std of the mean prediction I also tested the application of
the summary statistics approach to the computation of the across-subject mean
time course. This amounts to simply ignoring the stds of the subject-wise
predictions and estimating uncertainty exclusively from the mean predictions of
the subjects, i.e., you just compute the standard error of the average across
subjects. The resulting standard error is an order of magnitude or so larger
than the standard deviation of the mean prediction, but I'm undecided which one
to trust more. The summary statistics approach does not really apply here as
the number of trials varies across subjects and I cannot assume that the means
have equal variance. Also, the distribution of the mean prediction appears to
be a reasonable measure in the context of GP prediction. On the other hand,
larger uncertainty estimates usually don't hurt and the mean prediction may
conceptually be a questionable measure when you really do inference over a
hidden mean across subjects, not prediction.