response effects start earlier in ipsilateral primary motor cortex
I further wondered why I see relatively sharp motor effects aligned to the
response, but temporally smoothed and surprisingly early response effects for
first-dot-onset-aligned data (build-up in contralateral hemisphere from about
400 ms to a peak at 700 ms while the mode of the RT distribution is around 900
ms). The motoprep correlation analysis supports the sharp motor effect, but I
realised that this is only half of the story, because it only shows the average
signal to a contralateral choice. What happens in ipsilateral choices?

I have now extended the figure and show the average signals for both
contralteral and ipsilateral choices. For ipsilateral choices there is no sharp
peak close to the response. Rather, there is a shallow negative peak ca. 100 ms
before the response. The shape of the signal also suggests that there is a
reduction in signal in the ipsilateral hemisphere starting already around 700
to 600 ms before the response. Plotting the contra- and ipsilateral signals
together also highlights the comparably large effects in the contralateral
hemisphere which reases values around 0.2 normalised mean current units while
the drop in signal in the ipsilateral hemisphere remains above -0.1.

Seeing that the drop in signal in the ipsilateral hemisphere is rather
widespread, reaches similar or even larger amplitudes than the x-coordinate
signal (correlation_all_vs_early) and at least in the right hemisphere the
x-coordinate correlations occur mostly in the negative signal range, I worry
that the x-coordinate correlations are driven by the general drop in signal in
the ipsilateral hemisphere.