dot-specific effects contradict simple average rise or drop
Following up on my last commit (80013e0b) I checked whether the signal in
primary motor cortex only correlates with the dot x-coordinate, because there
is a general build-up or drop towards a particular choice and individual
x-coordinates are themselves correlated with the eventual choice. If this is
true, the x-coordinate correlations should become stronger as time passes
irrespective of which dot you look at. This is because at later time points
there is a larger difference between the left and right choices. The only thing
that could disturb this pattern of increasing correlations irrespective of dot
identity is that different dots provide different information about the final
choice. According to the choice_correlations figure the 4th dot indeed provides
less information about the final choice than other dots (smaller correlation
between x-coordinate and final choice) and the 5th dot provides a little bit
more information. This should lead to another modulation of the amount of
correlation between x-coordinate and source signals which rides on top of a
general increase: medium high correlations for the 3rd dot, low correlations
for the 4th dot and high correlations for the 5th dot.

As a singledot result for these 3 dots in source space already existed, I only
had to plot the corresponding time courses (x-coordinate correlation in area
4). Instead of a steady rise or drop in correlations the plots (left and right
hemisphere) show the pattern that I expected according to my previous results:
For the 3rd dot, correlations start to exceed permutation values from around
500 ms and drop back to baseline between 700 and 800 ms. This is consistent
with the sequential analysis where x correlations in primary motor cortex
appear between 300 and 500 ms after dot onset. Similarly, 4th and 5th dot x
correlations occur from about 600 and 700 ms, respectively. Only for the 4th
dot I also see a drop back to baseline (for the 5th dot the last time point for
which data is available corresponds to 490 ms after dot onset). The
correlations with the 4th dot might indeed be the smallest ones, but
correlations for 3rd and 5th dot appear to be at least of equal size (in the
right hemisphere correlations for the 3rd dot are much larger than those for
the fifth while in the right hemisphere correlations have similar magnitudes
with a short peak for the 5th dot exceeding the values of the 3rd).

Taken together the analysis of single dot correlations strongly suggests that
the x-coordinate correlations are not simply a shadow of general, motor-related
build-up or drop of signal.